them doctor upload cccd va cchn

+++ src/main/java/fpt/swp391/GlucoTrackAlert/dto/login/LoginResponse.java
@@ -8,9 +8,10 @@ import lombok.*;
 @AllArgsConstructor
 @Builder
 public class LoginResponse {
+
     private String token;
     private String email;
     private String role;
     private String message;
-    private Long doctorId;
+    private Integer doctorId;
 }
